REF: Refactor `nifreeze.data.pet.py` into an own module
Refactor `nifreeze.data.pet.py` into an own module:
- Improves encapsulation and logic.
- Reduces the complexity of the former `nifreeze.data.pet.py` module as
  functions are now spread across different modules with better limited
  purpose.
- Increases consistency with the approach adopted for the DWI data
  module.

Add a description of the module.
