+++ src/nifreeze/data/pet/base.py
@@ -24,23 +24,19 @@
 
 from __future__ import annotations
 
-import json
 from collections import namedtuple
-from collections.abc import Callable
 from pathlib import Path
-from typing import Any, Tuple
+from typing import Any
 
 import attrs
 import h5py
 import nibabel as nb
 import numpy as np
-from nibabel.spatialimages import SpatialImage
 from nitransforms.linear import Affine
 from nitransforms.resampling import apply
 from typing_extensions import Self
 
 from nifreeze.data.base import BaseDataset, _cmp, _data_repr, _has_ndim
-from nifreeze.utils.ndimage import get_data, load_api
 
 ATTRIBUTE_ABSENCE_ERROR_MSG = "PET '{attribute}' may not be None"
 """PET initialization array attribute absence error message."""
@@ -67,12 +63,6 @@ SCALAR_ATTRIBUTE_ERROR_MSG = (
     "PET '{attribute}' must be a numeric or single-element sequence object."
 )
 """PET initialization scalar attribute object error message."""
-
-TEMPORAL_FILE_KEY_ERROR_MSG = "{key} key not found in temporal file"
-"""PET temporal file key error message."""
-
-FRAME_TIME_START_KEY = "FrameTimesStart"
-"""PET frame time start key."""
 
 
 def format_scalar_like(value: Any, attr: attrs.Attribute) -> float:
@@ -389,156 +379,3 @@ class PET(BaseDataset[np.ndarray]):
             root = in_file["/0"]
             data = {k: np.asanyarray(v) for k, v in root.items() if not k.startswith("_")}
         return cls(**data)
-
-
-def from_nii(
-    filename: Path | str,
-    temporal_file: Path | str,
-    brainmask_file: Path | str | None = None,
-) -> PET:
-    """
-    Load PET data from NIfTI, creating a PET object with appropriate metadata.
-
-    Parameters
-    ----------
-    filename : :obj:`os.pathlike`
-        The NIfTI file.
-    temporal_file : :obj:`os.pathlike`
-        A JSON file containing temporal data. It must at least contain
-        ``FrameTimesStart`` data.
-    brainmask_file : :obj:`os.pathlike`, optional
-        A brainmask NIfTI file. If provided, will be loaded and
-        stored in the returned dataset.
-
-    Returns
-    -------
-    :obj:`~nifreeze.data.pet.PET`
-        A PET object storing the data, metadata, and any optional mask.
-
-    Raises
-    ------
-    :exc:`RuntimeError`
-        If ``FrameTimesStart`` is not provided (BIDS requires it).
-
-    """
-
-    filename = Path(filename)
-
-    # 1) Load a NIfTI
-    img = load_api(filename, SpatialImage)
-    fulldata = get_data(img)
-
-    # 2) Load the temporal data
-    with open(temporal_file, "r") as f:
-        temporal_attrs = json.load(f)
-
-    frame_time = temporal_attrs.get(FRAME_TIME_START_KEY, None)
-    if frame_time is None:
-        raise RuntimeError(TEMPORAL_FILE_KEY_ERROR_MSG.format(key=FRAME_TIME_START_KEY))
-
-    # 3) If a brainmask_file was provided, load it
-    brainmask_data = None
-    if brainmask_file is not None:
-        mask_img = load_api(brainmask_file, SpatialImage)
-        brainmask_data = np.asanyarray(mask_img.dataobj, dtype=bool)
-
-    # 4) Compute temporal attributes
-    midframe, total_duration = _compute_temporal_markers(np.asarray(frame_time))
-
-    # 5) Create and return the PET instance
-    return PET(
-        dataobj=fulldata,
-        affine=img.affine,
-        brainmask=brainmask_data,
-        midframe=midframe,
-        total_duration=total_duration,
-    )
-
-
-def _compute_temporal_markers(frame_time: np.ndarray) -> Tuple[np.ndarray, float]:
-    """Compute the frame temporal markers from the frame time values.
-
-    Computes the midframe times and the total duration following the principles
-    detailed below.
-
-    Let :math:`K` be the number of frames and :math:`t_{k}` be the :math:`k`-th
-    (start) frame time. For each frame :math:`k`, the frame duration
-    :math:`d_{k}` is defined as the difference between consecutive frame times:
-
-    .. math::
-       d_{k} = t_{k+1} - t_{k}
-
-    If necessary, the last frame duration is set to the value of the second to
-    last frame to match the appropriate dimensionality in this implementation.
-
-    Per-frame midpoints :math:`m_{k}` are computed as:
-
-    .. math::
-       m_{k} = t_{k} + \\frac{d_k}{2}
-
-    The total duration :math:`D` of the acquisition is a scalar computed as the
-    sum of the frame durations:
-
-    .. math::
-       D = \\sum_{k=1}^{K} d_{k}
-
-    or, equivalently, the difference between the last frame start and its
-    duration once the frame times have been time-origin shifted:
-
-    .. math::
-       D = t_{K} - d_{K}
-
-    Frame times are time-origin shifted (i.e. the earliest time is zeroed out)
-    if not already done at the beginning of the process for the sake of
-    simplicity.
-
-    Parameters
-    ----------
-    frame_time : :obj:`~numpy.ndarray`
-        Frame time values.
-
-    Returns
-    -------
-    :obj:`tuple`
-        Midpoint timing of each frame and total duration
-    """
-
-    # Time-origin shift: zero out the earliest time if necessary
-    # Flatten the array in case it is not a 1D array
-    if not np.isclose(frame_time.ravel()[0], 0):
-        frame_time -= frame_time.flat[0]
-
-    # If shape is e.g. (N,), then we can do
-    frame_duration = np.diff(frame_time)
-    if len(frame_duration) == (len(frame_time) - 1):
-        frame_duration = np.append(
-            frame_duration, frame_duration[-1]
-        )  # last frame same as second-last
-
-    midframe = frame_time + frame_duration / 2
-    total_duration = float(frame_time[-1] + frame_duration[-1])
-
-    return midframe, total_duration
-
-
-def compute_uptake_statistic(data: np.ndarray, stat_func: Callable[..., np.ndarray] = np.sum):
-    """Compute a statistic over all voxels for each frame on a PET sequence.
-
-    Assumes the last dimension corresponds to the number of frames in the
-    sequence.
-
-    Parameters
-    ----------
-    data : :obj:`~numpy.ndarray`
-        PET data.
-    stat_func : :obj:`~collections.abc.Callable`, optional
-        Function to apply over voxels (e.g., :func:`numpy.sum`,
-        :func:`numpy.mean`, :func:`numpy.std`)
-
-    Returns
-    -------
-    :obj:`~numpy.ndarray`
-        1D array of statistic values for each frame.
-    """
-
-    return stat_func(data.reshape(-1, data.shape[-1]), axis=0)
